Move some bus-mapping related errors out of eth-types.
