+++ bus-mapping/src/error.rs
@@ -1,0 +1,47 @@
+//! Error module for the bus-mapping crate
+
+use core::fmt::{Display, Formatter, Result as FmtResult};
+use eth_types::{Address, GethExecStep};
+use ethers_providers::ProviderError;
+use std::error::Error as StdError;
+
+/// Error type for any BusMapping related failure.
+#[derive(Debug)]
+pub enum Error {
+    /// Serde de/serialization error.
+    SerdeError(serde_json::error::Error),
+    /// Error while generating a trace.
+    TracingError,
+    /// JSON-RPC related error.
+    JSONRpcError(ProviderError),
+    /// OpcodeId is not a call type.
+    OpcodeIdNotCallType,
+    /// Account not found in the StateDB
+    AccountNotFound(Address),
+    /// Unable to figure out error at a [`GethExecStep`]
+    UnexpectedExecStepError(&'static str, Box<GethExecStep>),
+    /// Invalid [`GethExecStep`] due to an invalid/unexpected value in it.
+    InvalidGethExecStep(&'static str, Box<GethExecStep>),
+    /// Eth type related error.
+    EthTypeError(eth_types::Error),
+}
+
+impl From<eth_types::Error> for Error {
+    fn from(err: eth_types::Error) -> Self {
+        Error::EthTypeError(err)
+    }
+}
+
+impl From<ProviderError> for Error {
+    fn from(err: ProviderError) -> Self {
+        Error::JSONRpcError(err)
+    }
+}
+
+impl Display for Error {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl StdError for Error {}
